Wrap the 'OutputStream' that we give the user, to avoid overwriting.
Otherwise the user can keep the sink going directly to the connection
and start writing whenever she wants.

Note that the check is not thread-safe: the assumption (pervasive in
io-streams) related stuff is that only one thread at a time is working
on the Connection.  I plan to make this the case anyway in later
changes.